backend/requirements.txt
Django>=5.2.6
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0
python-docx>=0.8.11